Completed Expense page api request checking and connection with backend

+++ server/src/controllers/expense.controller.js
@@ -3,7 +3,10 @@ const expenseService = require("../services/expense.service");
 // Create Expense
 const createExpense = async (req, res) => {
   try {
-    const expense = await expenseService.createExpense(req.body);
+    const expense = await expenseService.createExpense({
+  ...req.body,
+  paidBy: req.user.id,
+});
 
     res.status(201).json({
       success: true,
@@ -20,7 +23,7 @@ const createExpense = async (req, res) => {
 // Get All Expenses
 const getAllExpenses = async (req, res) => {
   try {
-    const expenses = await expenseService.getAllExpenses();
+    const expenses = await expenseService.getAllExpenses(req.user.id);
 
     res.json({
       success: true,
@@ -61,14 +64,30 @@ const getExpenseById = async (req, res) => {
 // Update Expense
 const updateExpense = async (req, res) => {
   try {
-    const expense = await expenseService.updateExpense(
+    const expense = await expenseService.getExpenseById(req.params.id);
+
+    if (!expense) {
+      return res.status(404).json({
+        success: false,
+        message: "Expense not found",
+      });
+    }
+
+    if (expense.paidBy._id.toString() !== req.user.id) {
+      return res.status(403).json({
+        success: false,
+        message: "You can only modify your own expenses.",
+      });
+    }
+
+    const updatedExpense = await expenseService.updateExpense(
       req.params.id,
       req.body
     );
 
     res.json({
       success: true,
-      data: expense,
+      data: updatedExpense,
     });
   } catch (error) {
     res.status(500).json({
@@ -81,6 +100,21 @@ const updateExpense = async (req, res) => {
 // Delete Expense
 const deleteExpense = async (req, res) => {
   try {
+    const expense = await expenseService.getExpenseById(req.params.id);
+
+    if (!expense) {
+      return res.status(404).json({
+        success: false,
+        message: "Expense not found",
+      });
+    }
+
+    if (expense.paidBy._id.toString() !== req.user.id) {
+      return res.status(403).json({
+        success: false,
+        message: "You can only modify your own expenses.",
+      });
+    }
     await expenseService.deleteExpense(req.params.id);
 
     res.json({
